# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Khojsansar API",
        "description": "Khojsansar App API Documentation",
        "contact": {
            "email": "trishachhetri27@gmail.com"
        },
        "license": {
            "name": "Apache 2.4.58",
            "url": "https://www.apache.org/licenses/LICENSE-2.4.58.html"
        },
        "version": "3.0"
    },
    "paths": {
...
        "/api/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get list of facilities",
                "operationId": "313063116b188c2da61bebe23a871010",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Facility"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of Services",
                "operationId": "7cb57dfc3dc68b14a0f32724289e41ab",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Service"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facility": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "facility_name": {
                        "type": "string",
                        "example": "Wi-Fi"
                    },
                    "facility_logo": {
                        "type": "string",
                        "example": "wifi.png"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "facility_name": "Wi-Fi",
                    "facility_logo": "wifi.png"
                }
            },
...
            "Service": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service_name": {
                        "type": "string",
                        "example": "Buffet Service"
                    },
                    "service_logo": {
                        "type": "string",
                        "example": "buffet.png"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "service_name": "Buffet Service",
                    "service_logo": "buffet.png"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}